+++ buildSrc/src/main/kotlin/net/modgarden/barricade/gradle/Versions.kt
@@ -1,25 +1,28 @@
 package net.modgarden.barricade.gradle
 
 object Versions {
-    const val MOD = "1.0.0"
+    const val MOD = "1.1.0"
 
     const val MINECRAFT = "1.21.1"
     const val PARCHMENT_MINECRAFT = "1.21"
     const val NEOFORM = "$MINECRAFT-20240808.144430"
     const val PARCHMENT = "2024.07.07"
 
-    const val FABRIC_API = "0.102.0+1.21.1"
-    const val FABRIC_LOADER = "0.15.11"
-    const val MOD_MENU = "11.0.1"
+    const val FABRIC_API = "0.108.0+1.21.1"
+    const val FABRIC_LOADER = "0.16.9"
+    const val MOD_MENU = "11.0.3"
     const val FABRIC_LOADER_RANGE = ">=0.15"
     const val FABRIC_MINECRAFT_RANGE = ">=1.21"
 
-    const val NEOFORGE = "21.1.3"
+    const val NEOFORGE = "21.1.79"
     const val NEOFORGE_LOADER_RANGE = "[4,)"
     const val NEOFORGE_MINECRAFT_RANGE = "[1.21,)"
 
-    const val FABRIC_MIXIN = "0.15.0+mixin.0.8.7"
-    const val MIXIN_EXTRAS = "0.4.0"
+    const val FABRIC_MIXIN = "0.15.3+mixin.0.8.7"
+    const val MIXIN_EXTRAS = "0.4.1"
+
+    const val SODIUM = "mc1.21.1-0.6.0-beta.4"
+    const val FORGIFIED_FABRIC_API = "0.104.0+2.0.18+1.21.1"
 
     const val JAVA = 21
 }