+++ common/src/main/java/net/modgarden/barricade/client/model/OperatorUnbakedModel.java
@@ -16,7 +16,6 @@ import net.minecraft.client.resources.model.UnbakedModel;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.tags.TagKey;
 import net.minecraft.util.GsonHelper;
 import net.minecraft.world.item.Item;
 import net.modgarden.barricade.Barricade;
@@ -27,13 +26,13 @@ import java.lang.reflect.Type;
 import java.util.Collection;
 import java.util.function.Function;
 
-public class CreativeOnlyUnbakedModel implements UnbakedModel {
+public class OperatorUnbakedModel implements UnbakedModel {
     protected final UnbakedModel sourceModel;
-    protected final Either<TagKey<Item>, ResourceKey<Item>> requiredItem;
+    protected final Either<ResourceLocation, ResourceKey<Item>> requiredItem;
 
-    public CreativeOnlyUnbakedModel(
+    public OperatorUnbakedModel(
             UnbakedModel model,
-            Either<TagKey<Item>, ResourceKey<Item>> requiredItem) {
+            Either<ResourceLocation, ResourceKey<Item>> requiredItem) {
         this.sourceModel = model;
         this.requiredItem = requiredItem;
     }
@@ -54,10 +53,10 @@ public class CreativeOnlyUnbakedModel implements UnbakedModel {
         return BarricadeClient.getHelper().createCreativeOnlyModel(sourceModel.bake(modelBaker, textureGetter, modelState), requiredItem);
     }
 
-    public static class Deserializer implements JsonDeserializer<CreativeOnlyUnbakedModel> {
+    public static class Deserializer implements JsonDeserializer<OperatorUnbakedModel> {
         public static final ResourceLocation ID = Barricade.asResource("operator");
 
-        public CreativeOnlyUnbakedModel deserialize(JsonElement json, Type type, JsonDeserializationContext context) throws JsonParseException {
+        public OperatorUnbakedModel deserialize(JsonElement json, Type type, JsonDeserializationContext context) throws JsonParseException {
             if (!json.isJsonObject())
                 throw new JsonParseException("Cannot deserialize non JsonObject 'barricade:operator' model.");
 
@@ -68,16 +67,16 @@ public class CreativeOnlyUnbakedModel implements UnbakedModel {
             if (!jsonObject.has("required_item"))
                 throw new JsonParseException("Cannot get 'required_item' field from 'barricade:operator' model.");
 
-            Either<TagKey<Item>, ResourceKey<Item>> either = getEither(GsonHelper.getAsString(jsonObject, "required_item"));
+            Either<ResourceLocation, ResourceKey<Item>> either = getEither(GsonHelper.getAsString(jsonObject, "required_item"));
 
-            return new CreativeOnlyUnbakedModel(model, either);
+            return new OperatorUnbakedModel(model, either);
         }
 
-        private static Either<TagKey<Item>, ResourceKey<Item>> getEither(String id) throws JsonParseException {
-            Either<TagKey<Item>, ResourceKey<Item>> either;
+        private static Either<ResourceLocation, ResourceKey<Item>> getEither(String id) throws JsonParseException {
+            Either<ResourceLocation, ResourceKey<Item>> either;
             try {
                 if (id.startsWith("#"))
-                    either = Either.left(TagKey.create(Registries.ITEM, ResourceLocation.read(id.substring(1)).getOrThrow()));
+                    either = Either.left(ResourceLocation.read(id.substring(1)).getOrThrow());
                 else
                     either = Either.right(ResourceKey.create(Registries.ITEM, ResourceLocation.read(id).getOrThrow()));
             } catch (Exception ex) {