+++ common/src/main/java/net/modgarden/barricade/client/platform/BarricadeClientPlatformHelper.java
@@ -6,9 +6,8 @@ import net.minecraft.client.renderer.block.model.BlockModel;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.tags.TagKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
-import net.modgarden.barricade.client.util.OperatorItemPsuedoTag;
 
 import java.util.Collection;
 
@@ -18,6 +17,6 @@ public interface BarricadeClientPlatformHelper {
 
     BlockModel.Deserializer getBlockModelDeserializer();
 
-    BakedModel createCreativeOnlyModel(BakedModel model, Either<OperatorItemPsuedoTag, ResourceKey<Item>> requiredItem);
+    BakedModel createCreativeOnlyModel(BakedModel model, Either<ResourceLocation, ResourceKey<Item>> requiredItem);
 
 }