+++ common/src/main/java/net/modgarden/barricade/platform/BarricadePlatformHelper.java
@@ -1,0 +1,5 @@
+package net.modgarden.barricade.platform;
+
+public interface BarricadePlatformHelper {
+    boolean isModLoaded(String modId);
+}