+++ fabric/src/main/java/net/modgarden/barricade/fabric/client/BarricadeFabricClient.java
@@ -7,20 +7,34 @@ import com.google.gson.JsonParser;
 import com.mojang.datafixers.util.Pair;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientChunkEvents;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientWorldEvents;
 import net.fabricmc.fabric.api.client.model.loading.v1.PreparableModelLoadingPlugin;
 import net.fabricmc.fabric.api.client.rendering.v1.BuiltinItemRendererRegistry;
-import net.fabricmc.fabric.api.client.rendering.v1.InvalidateRenderStateCallback;
+import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderEvents;
+import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener;
+import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.minecraft.Util;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
+import net.minecraft.client.telemetry.events.WorldLoadEvent;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.PackType;
 import net.minecraft.server.packs.resources.Resource;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.util.GsonHelper;
+import net.minecraft.util.Unit;
+import net.minecraft.util.profiling.ProfilerFiller;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Blocks;
 import net.modgarden.barricade.Barricade;
 import net.modgarden.barricade.client.BarricadeClient;
-import net.modgarden.barricade.client.model.CreativeOnlyUnbakedModel;
+import net.modgarden.barricade.client.model.OperatorUnbakedModel;
+import net.modgarden.barricade.client.util.BarrierRenderUtils;
+import net.modgarden.barricade.client.util.OperatorItemPseudoTag;
 import net.modgarden.barricade.fabric.client.platform.BarricadeClientPlatformHelperFabric;
 import net.modgarden.barricade.client.renderer.block.AdvancedBarrierBlockRenderer;
 import net.modgarden.barricade.client.renderer.item.AdvancedBarrierItemRenderer;
@@ -37,8 +51,11 @@ import java.util.concurrent.Executor;
 import java.util.stream.Collectors;
 
 public class BarricadeFabricClient implements ClientModInitializer {
+    private static ItemStack lastItemInMainHand = ItemStack.EMPTY;
+    private static ItemStack lastItemInOffHand = ItemStack.EMPTY;
+
     private static final Gson GSON = new GsonBuilder()
-            .registerTypeAdapter(CreativeOnlyUnbakedModel.class, new CreativeOnlyUnbakedModel.Deserializer())
+            .registerTypeAdapter(OperatorUnbakedModel.class, new OperatorUnbakedModel.Deserializer())
             .create();
 
     @Override
@@ -55,8 +72,32 @@ public class BarricadeFabricClient implements ClientModInitializer {
             });
         });
 
-        InvalidateRenderStateCallback.EVENT.register(() -> {
+        ResourceManagerHelper.get(PackType.CLIENT_RESOURCES).registerReloadListener(new IdentifiableResourceReloadListener() {
+            private final OperatorItemPseudoTag.Loader listener = OperatorItemPseudoTag.Loader.INSTANCE;
 
+            @Override
+            public ResourceLocation getFabricId() {
+                return Barricade.asResource("operator_items");
+            }
+
+            @Override
+            public CompletableFuture<Void> reload(PreparationBarrier preparationBarrier, ResourceManager resourceManager, ProfilerFiller preparationsProfiler, ProfilerFiller reloadProfiler, Executor backgroundExecutor, Executor gameExecutor) {
+                return listener.reload(preparationBarrier, resourceManager, preparationsProfiler, reloadProfiler, backgroundExecutor, gameExecutor);
+            }
+        });
+
+        WorldRenderEvents.END.register(context -> {
+            LocalPlayer player = Minecraft.getInstance().player;
+            if (player == null)
+                return;
+            if (!ItemStack.isSameItemSameComponents(player.getMainHandItem(), lastItemInMainHand)) {
+                BarrierRenderUtils.refreshBarrierBlocks(player.getMainHandItem(), lastItemInMainHand);
+                lastItemInMainHand = player.getMainHandItem();
+            }
+            if (!ItemStack.isSameItemSameComponents(player.getOffhandItem(), lastItemInOffHand)) {
+                BarrierRenderUtils.refreshBarrierBlocks(player.getOffhandItem(), lastItemInOffHand);
+                lastItemInOffHand = player.getOffhandItem();
+            }
         });
 
         BlockRenderLayerMap.INSTANCE.putBlocks(RenderType.cutout(),
@@ -78,9 +119,9 @@ public class BarricadeFabricClient implements ClientModInitializer {
         );
     }
 
-    private static CompletableFuture<Map<ResourceLocation, CreativeOnlyUnbakedModel>> getCreativeUnbakedModels(ResourceManager manager, Executor executor) {
+    private static CompletableFuture<Map<ResourceLocation, OperatorUnbakedModel>> getCreativeUnbakedModels(ResourceManager manager, Executor executor) {
         return CompletableFuture.supplyAsync(() -> manager.listResources("models", fileName -> fileName.getPath().endsWith(".json")), executor).thenCompose(models -> {
-            ArrayList<CompletableFuture<Pair<ResourceLocation, CreativeOnlyUnbakedModel>>> creativeModels = new ArrayList<>();
+            ArrayList<CompletableFuture<Pair<ResourceLocation, OperatorUnbakedModel>>> creativeModels = new ArrayList<>();
             for (Map.Entry<ResourceLocation, Resource> resource : models.entrySet()) {
                 creativeModels.add(CompletableFuture.supplyAsync(() -> {
                     try {
@@ -91,10 +132,10 @@ public class BarricadeFabricClient implements ClientModInitializer {
                         if (!element.isJsonObject() || !element.getAsJsonObject().has("loader"))
                             return null;
                         String loaderKey = GsonHelper.getAsString(element.getAsJsonObject(), "loader");
-                        if (!loaderKey.equals(CreativeOnlyUnbakedModel.Deserializer.ID.toString()))
+                        if (!loaderKey.equals(OperatorUnbakedModel.Deserializer.ID.toString()))
                             return null;
 
-                        return Pair.of(resourceLocation, GSON.fromJson(element, CreativeOnlyUnbakedModel.class));
+                        return Pair.of(resourceLocation, GSON.fromJson(element, OperatorUnbakedModel.class));
                     } catch (Exception ex) {
                         Barricade.LOG.error("Failed to load 'barricade:creative_mode' model", ex);
                     }