+++ fabric/src/main/java/net/modgarden/barricade/fabric/client/platform/BarricadeClientPlatformHelperFabric.java
@@ -6,10 +6,9 @@ import net.minecraft.client.renderer.block.model.BlockModel;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.tags.TagKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.modgarden.barricade.client.platform.BarricadeClientPlatformHelper;
-import net.modgarden.barricade.client.util.OperatorItemPsuedoTag;
 import net.modgarden.barricade.fabric.client.model.CreativeOnlyBakedModel;
 
 import java.util.Collection;
@@ -26,7 +25,7 @@ public class BarricadeClientPlatformHelperFabric implements BarricadeClientPlatf
     }
 
     @Override
-    public BakedModel createCreativeOnlyModel(BakedModel model, Either<OperatorItemPsuedoTag, ResourceKey<Item>> requiredItem) {
+    public BakedModel createCreativeOnlyModel(BakedModel model, Either<ResourceLocation, ResourceKey<Item>> requiredItem) {
         return new CreativeOnlyBakedModel(model, requiredItem);
     }