+++ neoforge/src/main/java/net/modgarden/barricade/neoforge/client/BarricadeNeoForgeClient.java
@@ -3,10 +3,17 @@ package net.modgarden.barricade.neoforge.client;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
+import net.minecraft.util.Unit;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.modgarden.barricade.Barricade;
 import net.modgarden.barricade.client.BarricadeClient;
-import net.modgarden.barricade.client.model.CreativeOnlyUnbakedModel;
+import net.modgarden.barricade.client.model.OperatorUnbakedModel;
+import net.modgarden.barricade.client.util.BarrierRenderUtils;
+import net.modgarden.barricade.client.util.OperatorItemPseudoTag;
 import net.modgarden.barricade.neoforge.client.bewlr.BarricadeBEWLR;
 import net.modgarden.barricade.client.renderer.block.AdvancedBarrierBlockRenderer;
 import net.modgarden.barricade.neoforge.client.model.CreativeOnlyUnbakedModelGeometry;
@@ -20,14 +27,40 @@ import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.client.event.EntityRenderersEvent;
 import net.neoforged.neoforge.client.event.ModelEvent;
+import net.neoforged.neoforge.client.event.RegisterClientReloadListenersEvent;
+import net.neoforged.neoforge.client.event.RenderLevelStageEvent;
 import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 import net.neoforged.neoforge.client.extensions.common.RegisterClientExtensionsEvent;
 import net.neoforged.neoforge.client.model.geometry.IGeometryLoader;
+import net.neoforged.neoforge.event.level.ChunkEvent;
 
 @Mod(value = Barricade.MOD_ID, dist = Dist.CLIENT)
 public class BarricadeNeoForgeClient {
     public BarricadeNeoForgeClient(IEventBus eventBus) {
         BarricadeClient.init(new BarricadeClientPlatformHelperNeoForge());
+    }
+
+    @EventBusSubscriber(bus = EventBusSubscriber.Bus.GAME, modid = Barricade.MOD_ID, value = Dist.CLIENT)
+    public static class GameEvents {
+        private static ItemStack lastItemInMainHand = ItemStack.EMPTY;
+        private static ItemStack lastItemInOffHand = ItemStack.EMPTY;
+
+        @SubscribeEvent
+        public static void onRenderLevel(RenderLevelStageEvent event) {
+            if (event.getStage() == RenderLevelStageEvent.Stage.AFTER_LEVEL) {
+                LocalPlayer player = Minecraft.getInstance().player;
+                if (player == null)
+                    return;
+                if (!ItemStack.isSameItemSameComponents(player.getMainHandItem(), lastItemInMainHand)) {
+                    BarrierRenderUtils.refreshBarrierBlocks(player.getMainHandItem(), lastItemInMainHand);
+                    lastItemInMainHand = player.getMainHandItem();
+                }
+                if (!ItemStack.isSameItemSameComponents(player.getOffhandItem(), lastItemInOffHand)) {
+                    BarrierRenderUtils.refreshBarrierBlocks(player.getOffhandItem(), lastItemInOffHand);
+                    lastItemInOffHand = player.getOffhandItem();
+                }
+            }
+        }
     }
 
     @EventBusSubscriber(modid = Barricade.MOD_ID, bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
@@ -39,7 +72,12 @@ public class BarricadeNeoForgeClient {
 
         @SubscribeEvent
         public static void registerGeometryLoaders(ModelEvent.RegisterGeometryLoaders event) {
-            event.register(CreativeOnlyUnbakedModel.Deserializer.ID, new CreativeOnlyDeserializer());
+            event.register(OperatorUnbakedModel.Deserializer.ID, new CreativeOnlyDeserializer());
+        }
+
+        @SubscribeEvent
+        public static void registerReloadListener(RegisterClientReloadListenersEvent event) {
+            event.registerReloadListener(OperatorItemPseudoTag.Loader.INSTANCE);
         }
 
         @SubscribeEvent
@@ -54,11 +92,11 @@ public class BarricadeNeoForgeClient {
     }
 
     public static class CreativeOnlyDeserializer implements IGeometryLoader<CreativeOnlyUnbakedModelGeometry> {
-        private static final CreativeOnlyUnbakedModel.Deserializer DESERIALIZER = new CreativeOnlyUnbakedModel.Deserializer();
+        private static final OperatorUnbakedModel.Deserializer DESERIALIZER = new OperatorUnbakedModel.Deserializer();
 
         @Override
         public CreativeOnlyUnbakedModelGeometry read(JsonObject json, JsonDeserializationContext context) throws JsonParseException {
-            return new CreativeOnlyUnbakedModelGeometry(DESERIALIZER.deserialize(json, CreativeOnlyUnbakedModel.class, context));
+            return new CreativeOnlyUnbakedModelGeometry(DESERIALIZER.deserialize(json, OperatorUnbakedModel.class, context));
         }
     }
 }