+++ neoforge/src/main/java/net/modgarden/barricade/neoforge/client/platform/BarricadeClientPlatformHelperNeoForge.java
@@ -6,10 +6,10 @@ import net.minecraft.client.renderer.block.model.BlockModel;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.tags.TagKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.modgarden.barricade.client.platform.BarricadeClientPlatformHelper;
-import net.modgarden.barricade.client.util.OperatorItemPsuedoTag;
+import net.modgarden.barricade.client.util.OperatorItemPseudoTag;
 import net.modgarden.barricade.neoforge.client.model.CreativeOnlyBakedModel;
 import net.neoforged.neoforge.client.model.ExtendedBlockModelDeserializer;
 
@@ -27,8 +27,8 @@ public class BarricadeClientPlatformHelperNeoForge implements BarricadeClientPla
     }
 
     @Override
-    public BakedModel createCreativeOnlyModel(BakedModel model, Either<OperatorItemPsuedoTag, ResourceKey<Item>> tagOrBlock) {
-        return new CreativeOnlyBakedModel(model, tagOrBlock);
+    public BakedModel createCreativeOnlyModel(BakedModel model, Either<ResourceLocation, ResourceKey<Item>> requiredItem) {
+        return new CreativeOnlyBakedModel(model, requiredItem);
     }
 
 }